Trying to synchronize the switches does not work.
The ControlsStatusSettings class was created to save the switch status and be able to synchronize them, but it was not achieved because the switch on one screen does not take the status of the other switch.

It is hoped to protect the code to continue with something else and then return to this topic.