Changes to spinners and status bar.
"128,000" was set as the default value for the bit rate spinner. For this reason, the spinners also had to be enlarged.
The color of the top status bar was changed to black, as is the UX design.

It is expected that all spinners will have a default value and the same design.