Changed the "Start" recording button.
The ToggleButton initialization code was moved from the "MainActiviy" class to the "HomeFragment" class.
Also, some methods of the "MainActiviy" class were made public so that everything does not have to be moved to the "HomeFragment" class.

The "Start" recording button is expected to work even if you change menus and return to the home page.