Disable menu and switch.
Code was added to disable the bottom menu and the switches on the home screen, so that you can't switch screens and change the recording settings.
The code to enable it again doesn't work.

It is expected to protect the code done so far and then we will review what is not working with the enabling of the components.