Changes to sampling rate.
Added new list of sampling rate values ​​for motion and GPS.
Modified the "Sample Rate" spinners for motion and GPS.
Modified where the "Sample Rate" value is passed for motion and audio sensors.

It is expected that with the implemented changes the logic defined for the application will be respected.