Modification of the info.txt file
The configuration values ​​were added to the info.txt file
And the code for saving and retrieving the state of the switches and spinners was optimized.

It is expected that the info.txt file will show the configuration that was in place at the time of the experiment.
As well as optimizing the code for saving and retrieving the configuration.