Modification to maintain the screen state.
A setting was added to the "AndroidManifiest.xml" to maintain the screen state when switching to the background.

It is expected that, when starting the recording, switching the application to the background and then pressing on the notification, when the application is opened it will show the stop button and the locked options.